+++ collada_parser/CHANGELOG.rst
@@ -2,6 +2,20 @@
 Changelog for package collada_parser
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.10.20 (2014-08-01)
+--------------------
+* Added extract actuators, fix for using nominal torque
+* Removed visual and collision meshes if there are no vertices
+* Removed use of visual and collision groups
+* Fix problem with root coordinates
+* Contributors: YoheiKakiuchi
+
+1.10.19 (2014-02-15)
+-------------------
+* fix, joint axis should be rotated depend on local coords
+* fix overwriting velocity limit
+* Contributors: YoheiKakiuchi
+
 1.10.18 (2013-12-04)
 --------------------
 * add DEPENDS for kdl_parser